requirements.txt
pyyaml
torch
torchvision
meshcat
pyquaternion
numpy
opencv-python
matplotlib
scipy
gym